@@ -1,6 +1,6 @@
 class ApiController < ActionController::Base
   protect_from_forgery with: :null_session
-  skip_before_filter :verify_authenticity_token
+  skip_before_action :verify_authenticity_token
 
   before_action :destroy_session
   before_action :set_default_response_format
