@@ -1,4 +1,4 @@
-class UserLanguage < ActiveRecord::Base
+class UserLanguage < ApplicationRecord
   belongs_to :user
   belongs_to :language
 
