@@ -2,18 +2,18 @@ require 'rails_helper'
 
 describe Api::V1::RegistrationsController do
 
-  let(:headers) { {HTTP_ACCEPT: 'application/json'} }
+  let(:headers) { {'HTTP_ACCEPT': 'application/json'} }
 
   describe 'POST /api/v1/users/' do
 
 
     describe 'register a user' do
       it 'with valid sign up returns user & token' do
-        post '/api/v1/users', {user:{user_name: 'Thomas',
+        post '/api/v1/users', params: {user:{user_name: 'Thomas',
                                email: 'thomas@craft.com',
                                gender: 'Male',
                                password: 'password',
-                               password_confirmation: 'password'}}, headers
+                               password_confirmation: 'password'}}, headers: headers
 
         expect(response_json['message']).to eq('success')
         expect(response_json['user']['user_name']).to eq('Thomas')
@@ -24,32 +24,32 @@ describe Api::V1::RegistrationsController do
       end
 
       it 'with an invalid password confirmation returns error message' do
-        post '/api/v1/users', {user:{user_name: 'Thomas',
+        post '/api/v1/users', params: {user:{user_name: 'Thomas',
                                      email: 'thomas@craft.com',
                                      gender: 'Male',
                                      password: 'password',
-                                     password_confirmation: 'wrong_password'}}, headers
+                                     password_confirmation: 'wrong_password'}}, headers: headers
         expect(response_json['errors']['password_confirmation']).to eq(['doesn\'t match Password'])
         expect(response.status).to eq 422
       end
 
       it 'with an invalid email returns error message' do
-        post '/api/v1/users', {user:{user_name: 'Thomas',
-                                     email: 'thomas@craft',
+        post '/api/v1/users', params: {user:{user_name: 'Thomas',
+                                     email: 'thomas@cr aft',
                                      gender: 'Male',
                                      password: 'password',
-                                     password_confirmation: 'password'}}, headers
+                                     password_confirmation: 'password'}}, headers: headers
         expect(response_json['errors']['email']).to eq(['is invalid'])
         expect(response.status).to eq 422
       end
 
       it 'with an registered email returns error message' do
         FactoryGirl.create(:user, email: 'thomas@craft.com')
-        post '/api/v1/users', {user:{user_name: 'Thomas',
+        post '/api/v1/users', params: {user:{user_name: 'Thomas',
                                      email: 'thomas@craft.com',
                                      gender: 'Male',
                                      password: 'password',
-                                     password_confirmation: 'password'}}, headers
+                                     password_confirmation: 'password'}}, headers: headers
         expect(response_json['errors']['email']).to eq(['has already been taken'])
         expect(response.status).to eq 422
       end
@@ -63,7 +63,7 @@ describe Api::V1::RegistrationsController do
         end
 
         it 'allows user to register with valid authorization' do
-          post '/api/v1/users/auth/facebook/callback', {}, headers
+          post '/api/v1/users/auth/facebook/callback', params: nil, headers: headers
           expect(response_json['message']).to eq('success')
           expect(response_json['user']['user_name']).to eq('Thomas Ochman')
           expect(response_json['user']['token']).to_not be nil
@@ -72,7 +72,7 @@ describe Api::V1::RegistrationsController do
 
         it 'fails to register user with invalid authorization' do
           OmniAuth.config.mock_auth[:facebook] = :invalid_credentials
-          post '/api/v1/users/auth/facebook/callback', {}, headers
+          post '/api/v1/users/auth/facebook/callback', params: nil, headers: headers
           expect(response_json['errors']).to eq('authentication error')
           expect(response.status).to eq 401
         end
